Cast the time_t variables as long int, since that's what %ld expects.

+++ src/analysisd/alerts/log.c
@@ -184,7 +184,7 @@ void OS_LogOutput(Eventinfo *lf)
         "** Alert %ld.%ld:%s - %s\n"
         "%d %s %02d %s %s%s%s\nRule: %d (level %d) -> '%s'"
         "%s%s%s%s%s%s%s%s%s%s%s%s%s%s\n%.1256s\n",
-        lf->time,
+        (long int)lf->time,
         __crt_ftell,
         lf->generated_rule->alert_opts & DO_MAILALERT ? " mail " : "",
         lf->generated_rule->group,
@@ -269,7 +269,7 @@ void OS_Log(Eventinfo *lf)
             "** Alert %ld.%ld:%s - %s\n"
             "%d %s %02d %s %s%s%s\nRule: %d (level %d) -> '%s'"
             "%s%s%s%s%s%s%s%s%s%s%s%s%s%s\n%.1256s\n",
-            lf->time,
+            (long int)lf->time,
             __crt_ftell,
             lf->generated_rule->alert_opts & DO_MAILALERT ? " mail " : "",
             lf->generated_rule->group,
@@ -342,7 +342,7 @@ void OS_CustomLog(const Eventinfo *lf, const char *format)
     /* Replace all the tokens */
     os_strdup(format, log);
 
-    snprintf(tmp_buffer, 1024, "%ld", lf->time);
+    snprintf(tmp_buffer, 1024, "%ld", (long int)lf->time);
     tmp_log = searchAndReplace(log, CustomAlertTokenName[CUSTOM_ALERT_TOKEN_TIMESTAMP], tmp_buffer);
     if (log) {
         os_free(log);
